requirements.txt
flask
gunicorn
openpyxl
pandas
pypdf
python-docx
